requirements.txt
# Core RAG dependencies
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-pinecone>=0.2.0
langchain-text-splitters>=0.3.0

# Vector store
pinecone-client>=3.0.0

# LLM / embedding backend
huggingface-hub>=0.24.0
sentence-transformers>=3.0.0
torch>=2.0.0

# PDF loading
pypdf>=4.0.0

# Web UI
gradio>=4.0.0

# Config
python-dotenv>=1.0.0

# Dev / test
pytest>=8.0.0
pytest-cov>=5.0.0
